requirements.txt
numpy==1.18.*
pandas==1.1.*
tensorflow==2.3.*
scikit-learn==0.23.*
scikit-optimize==0.8.*
sklearn-genetic==0.3.*
matplotlib==3.3.*
pdoc3==0.9.*
gitpython==3.1.*
pydot==1.4.*